refactored ContactInfo section
